Refactored class and adjusted tests
